requirements.txt
iniconfig==2.0.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.4
